chore: move cache into its own file; into libs/ folder along with config.js
